+++ sparkstreaming/src/main/scala/com/griddynamics/stopbot/spark/stream/KafkaStructureToParquet.scala
@@ -66,7 +66,7 @@ object KafkaStructureToParquet extends App {
       watermark = appConf.getDuration("spark.watermark"),
       minEvents = appConf.getLong("app.min-events"),
       maxEvents = appConf.getLong("app.max-events"),
-      minRate = appConf.getLong("app.min-rate")
+      minRate = appConf.getDouble("app.min-rate")
     )
 
   val parquet =
